Move EncodedPatternItem to API to avoid large amounts of copy-paste code for custom patterns in addons
