+++ src/main/java/appeng/crafting/pattern/CraftingPatternItem.java
@@ -15,6 +15,7 @@ import net.minecraft.world.item.crafting.RecipeManager;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.Level;
 
+import appeng.api.crafting.EncodedPatternItem;
 import appeng.api.stacks.AEItemKey;
 import appeng.core.AELog;
 import appeng.menu.AutoCraftingMenu;
